:sparkles: feat : mypage Detail Button
feat : mypage Detail Button
