+++ src/@components/myPage/style.tsx
@@ -11,3 +11,33 @@ export const MainLogoWrapper = styled.div`
     margin-top: 30px;
     
 `
+export const DetailWrapper = styled.div`
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+
+    margin-top: 32px;
+    
+`
+export const DetailProfileBox = styled.div`
+    
+    height: 154px;
+    flex-grow: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    background-color: #F8F8FA;
+    border-radius: 10px;
+    /* 두번째 자식만 */
+    &:nth-child(2){
+        margin: 0px 20px;
+    }
+    
+
+`
+export const DetailText = styled.h1`
+    margin-top: 16px;
+    size: 20px;
+    font-weight: bold;
+`
