+++ src/@pages/review/style.ts
@@ -1,0 +1,30 @@
+import { styled } from "styled-components";
+
+export const ReviewHeader = styled.h1`
+display: flex;
+    font-size: 36px;
+    font-weight: 700;
+    text-align: center;
+    margin-top:150px;
+    margin-bottom: 50px;
+`
+export const ReviewHeaderWrapper = styled.div`
+    display: flex;
+    flex-direction: column;
+    
+    align-items: center;
+`
+export const ReviewForm = styled.form`
+display: flex;
+    flex-direction: column;
+    width: 300px;
+    
+`
+export const ReviewInput = styled.input`
+border:  1px solid black;
+margin-bottom: 10px;
+`
+export const ReviewButton = styled.button`
+margin-top: 30px;
+border:  1px solid black;
+`
